Fix: Rename components to follow conventions
