+++ src/components/BubbleChat/GroupChatBubble/index.tsx
@@ -2,7 +2,7 @@ import { GroupMessage } from '@/model/GroupMessageModel'
 import { Flex } from 'antd'
 import Cookies from 'js-cookie'
 import React, { ReactNode } from 'react'
-import ProfileContact from '../../profile'
+import ProfileContact from '../../Profile'
 import SentAt from '../components/sendAt'
 
 export interface GroupChatBubbleProps extends GroupMessage {
