+++ src/components/MenuInfo/components/ButtonDeleteGroup/index.tsx
