+++ src/components/MenuInfo/components/ButtonRemoveMember/index.tsx
