+++ src/components/MenuInfo/components/buttonremove/index.tsx
