+++ src/components/MenuInfo/components/modalAlert/index.tsx
