+++ src/components/MenuInfo/components/modalAlert/style/style.tsx
