+++ src/components/MenuInfo/index.tsx
