+++ src/components/MenuInfo/style/style.css
