+++ src/components/MenuInfo/style/style.tsx
