+++ src/components/Profile/index.tsx
@@ -3,8 +3,8 @@ import { ChatContext } from '@/contexts/chatContext'
 import { useGetUsersAndFriends } from '@/utils/hooks/useGetUsersAndFriends'
 import { Flex } from 'antd'
 import { useContext } from 'react'
-import CargoProfile from './cargo'
-import ImageProfile from './imageProfile'
+import CargoProfile from './Cargo'
+import ProfileImage from './ProfileImage'
 import NameProfile from './ProfileName'
 
 interface ProfileContactProps {
@@ -23,7 +23,7 @@ const ProfileContact = ({ fromUserId }: ProfileContactProps) => {
 
   return (
     <Flex key={fromUser?.id} align="center" gap="0.5rem">
-      <ImageProfile size="2.5rem" image={fromUser?.avatar || defaultAvatar} />
+      <ProfileImage size="2.5rem" image={fromUser?.avatar || defaultAvatar} />
       <Flex vertical align="start">
         <CargoProfile cargo={'member'} />
         <NameProfile colortext="#FFF" user={fromUser?.nome} />
