+++ src/layouts/layout.tsx
@@ -2,8 +2,8 @@ import { Chat } from '@/components/Chat'
 import { ChatContainer } from '@/components/ChatContainer'
 import { InputBar } from '@/components/InputBar'
 import { Aside } from '@/components/Aside'
-import HeaderChat from '@/components/header'
-import MenuGroup from '@/components/menuGroup'
+import HeaderChat from '@/components/Header'
+import MenuGroup from '@/components/MenuGroup'
 import { Flex } from 'antd'
 import Cookies from 'js-cookie'
 import { useState } from 'react'
